Move filter converter to separate class
